Перенёс ненужную инициализацию аттрибутов картинки в ProfileFragment в XML файл Убрал забытые комменты Создал базовый класс для OnPageChangeListener Разделил зачем-то ViewHolder с RecyclerView для списков Убрал избыточные параметры в Interactor'ах
