+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/SplashActivity.kt
@@ -21,6 +21,6 @@ class SplashActivity : AppCompatActivity() {
         Handler().postDelayed({
             val intent = Intent(this@SplashActivity, MainActivity::class.java)
             startActivity(intent)
-        },0)
+        },2)
     }
 }
