+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/MainActivity.kt
@@ -87,19 +87,6 @@ class MainActivity : MvpAppCompatActivity(), MainMvpViewInterface {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-//TODO: Если не нужно - удаляй
-//        imageToolbarSearch.setOnClickListener {
-//            presenter.findContent(edittextToolbarSearch.text as String?)
-//        }
-//        edittextToolbarSearch.setOnEditorActionListener { editText, actionId, event ->
-//            if (actionId == EditorInfo.IME_ACTION_DONE) {
-//                presenter.findContent(editText.text.toString())
-//                true
-//            } else {
-//                false
-//            }
-//        }
-
         bottomNavBar.enableAnimation(false)
         bottomNavBar.onNavigationItemSelectedListener = BottomNavigationView.OnNavigationItemSelectedListener { item ->
             if (item.isChecked) {
