+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/profile/ProfileFragment.kt
@@ -25,12 +25,8 @@ class ProfileFragment : MvpAppCompatFragment(), ProfileMvpView {
     override fun fillProfileScreen(profile: UserProfile) {
         ImageProvider.loadImage(profile.pictureUrl, fragmentView.imageUserScreenPhoto)
         fragmentView.apply {
-            imageUserScreenPhoto.apply {
-                adjustViewBounds = true
-                scaleType = ImageView.ScaleType.FIT_XY
-                setOnClickListener {
-                    EditPhotoDialog().show(fragmentManager, EditPhotoDialog.TAG)
-                }
+            imageUserScreenPhoto.setOnClickListener {
+                EditPhotoDialog().show(fragmentManager, EditPhotoDialog.TAG)
             }
             textUserScreenName.text = profile.name
             textUserScreenBirth.text = profile.birthDate
