+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/search/BaseOnTabChangeListener.kt
@@ -1,0 +1,11 @@
+package ru.shcherbakovDmitry.ss.androidTraineeEducation.mainscreen.search
+
+import android.support.v4.view.ViewPager
+
+open class BaseOnTabChangeListener : ViewPager.OnPageChangeListener {
+    override fun onPageScrollStateChanged(state: Int) {}
+
+    override fun onPageScrolled(position: Int, positionOffset: Float, positionOffsetPixels: Int) {}
+
+    override fun onPageSelected(position: Int) {}
+}
