+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/search/SearchFieldObserver.kt
@@ -1,5 +1,5 @@
 package ru.shcherbakovDmitry.ss.androidTraineeEducation.mainscreen.search
 
 interface SearchFieldObserver {
-    fun requestContent(key: String?)
+    fun requestContent()
 }
