+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/search/charityeventtab/EventListAdapter.kt
@@ -11,62 +11,16 @@ import ru.shcherbakovDmitry.ss.androidTraineeEducation.utilites.ImageProvider
 import ru.shcherbakovDmitry.ss.androidTraineeEducation.mainscreen.dataclasses.CharityEvent
 import java.util.*
 
-class EventListAdapter(private val eventArray: Array<CharityEvent>) : RecyclerView.Adapter<EventListAdapter.ViewHolder>() {
+class EventListAdapter(private val eventArray: Array<CharityEvent>) : RecyclerView.Adapter<EventViewHolder>() {
 
-    class ViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
-
-        private fun getLocaleQuantityString(id: Int, quantity: Int): String {
-            val configuration = Configuration(view.context!!.resources.configuration).apply {
-                setLocale(Locale("ru"))
-            }
-            return view.context!!.createConfigurationContext(configuration).resources.getQuantityString(id, quantity, quantity)
-        }
-
-        fun bind(event: CharityEvent) {
-            view.apply {
-                titleEventItem.text = event.title
-                textEventItemDescription.text = event.description
-                val startDateString = event.startDate.split(".")
-                val startDate = Calendar.getInstance().apply {
-                    set(Calendar.YEAR, startDateString[2].toInt())
-                    set(Calendar.MONTH, startDateString[1].toInt())
-                    set(Calendar.DAY_OF_MONTH, startDateString[0].toInt())
-                }
-                val endDateString = event.endDate.split(".")
-                val endDate = Calendar.getInstance().apply {
-                    set(Calendar.YEAR, endDateString[2].toInt())
-                    set(Calendar.MONTH, endDateString[1].toInt())
-                    set(Calendar.DAY_OF_MONTH, endDateString[0].toInt())
-                }
-                var dayDifference = endDate.get(Calendar.DAY_OF_YEAR) - startDate.get(Calendar.DAY_OF_YEAR)
-                if (endDate.get(Calendar.DAY_OF_YEAR) < startDate.get(Calendar.DAY_OF_YEAR)) {
-                    dayDifference += if (startDate.get(Calendar.YEAR) % 4 == 0) {
-                        366
-                    } else {
-                        365
-                    }
-                }
-                textEventItemDate.text = String.format(resources.getString(R.string.event_date_expiration_info),
-                        getLocaleQuantityString(R.plurals.event_date_expiration, dayDifference),
-                        startDateString[0].toInt(),
-                        startDateString[1].toInt(),
-                        endDateString[0].toInt(),
-                        endDateString[1].toInt()
-                )
-                ImageProvider.loadImage(event.pictureURL,imageEventItem)
-            }
-        }
-
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, state: Int): ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, state: Int): EventViewHolder {
         val view = LayoutInflater.from(parent.context)
                 .inflate(R.layout.item_event_list, parent, false)
-        return ViewHolder(view)
+        return EventViewHolder(view)
     }
 
     override fun getItemCount() = eventArray.size
 
-    override fun onBindViewHolder(viewHolder: ViewHolder, item: Int) = viewHolder.bind(eventArray[item])
+    override fun onBindViewHolder(viewHolder: EventViewHolder, item: Int) = viewHolder.bind(eventArray[item])
 
 }
