+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/search/organisationtab/OrganisationTabFragment.kt
@@ -48,7 +48,7 @@ class OrganisationTabFragment : MvpAppCompatFragment(), OrganisationTabMvpView {
             val divider = DividerItemDecoration(context, LinearLayoutManager.VERTICAL)
             addItemDecoration(divider)
         }
-        presenter.requestContent("")
+        presenter.requestContent()
         return fragmentView
     }
