+++ app/src/main/res/layout/fragment_profile_screen.xml
@@ -18,7 +18,9 @@
                 android:id="@+id/imageUserScreenPhoto"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                tools:src="@drawable/image_user_placeholder" />
+                tools:src="@drawable/image_user_placeholder"
+                android:adjustViewBounds="true"
+                android:scaleType="fitXY"/>
 
             <LinearLayout
                 android:id="@+id/layoutUserScreenInfo"
